Fixed GCAlignment post call
It now sends the sequences from the frontend as a stringified array
then receives and parses it into a list in the backend.

Sequences are now sent and received properly as a list. It's also
written in a way that will make it usable if we want to compare
more sequences in the future.